requirements.txt
Django==1.6.7
Markdown==2.5
MySQL-python==1.2.5
django-crispy-forms==1.4.0
django-filter==0.7
django-xadmin==0.5.0
djangorestframework==2.4.2
mysqlclient==1.3.3
virtualenv==1.11.6
